+++ APLesson_03/ex_03.java
@@ -6,7 +6,7 @@ public class ex_03{
 		double avgBanana = 7.5;
 		
 		System.out.println("How tall are you (in inches)?");
-		double height = bananaScale.nextInt();
+		double height = bananaScale.nextDouble();
 		
 		System.out.println("You are approximately " + height/avgBanana + " bananas tall.");
 	}
